feat: 프록시와 연관관계 실습 코드 추가
