+++ Eunji/week5/jpa-mapping/src/main/java/mapping/Member.java
@@ -2,51 +2,18 @@ package mapping;
 
 import jakarta.persistence.*;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Entity
 public class Member {
     @Id @GeneratedValue
     @Column(name = "MEMBER_ID")
-    private Long id;
+    private Long id; //기본키
 
+    @Column(name = "MEMBER_NAME")
     private String name;
 
-    private String city;
-
-    private String street;
-
-    private String zipcode;
-
-    //orders 선언
-    @OneToMany(mappedBy = "member")
-    private List<Order> orders = new ArrayList<>();
-
-    public Member() {}
-
-    public Member(String name, String city, String street, String zipcode) {
-        this.name = name;
-        this.city = city;
-        this.street = street;
-        this.zipcode = zipcode;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-
-    public List<Order> getOrders() {
-        return orders;
-    }
-
-    public void setOrders(List<Order> orders) {
-        this.orders = orders;
-    }
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "TEAM_ID") // 연관관계 주인 -> 진짜 주인(매핑)
+    private Team team;
 
     public Long getId() {
         return id;
@@ -56,27 +23,18 @@ public class Member {
         this.id = id;
     }
 
-    public String getName() {
-        return name;
-    }
 
     public void setName(String name) {
         this.name = name;
     }
 
-    public String getCity() {
-        return city;
+
+
+    public Team getTeam() {
+        return team;
     }
 
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getZipcode() {
-        return zipcode;
-    }
-
-    public void setZipcode(String zipcode) {
-        this.zipcode = zipcode;
+    public void setTeam(Team team) {
+        this.team = team;
     }
 }
