--- Eunji/week5/jpa-mapping/src/main/java/mapping/Order.java
@@ -1,74 +1,0 @@
-package mapping;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name="ORDERS")
-public class Order {
-    @Id @GeneratedValue
-    @Column(name = "ORDER_ID")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "MEMBER_ID") //외래키 매핑
-    private Member member;
-
-    //order items
-    @OneToMany(mappedBy = "order")
-    private List <OrderItem> orderItems = new ArrayList<>();
-
-    private LocalDateTime orderDate;
-
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    //연관관계 편의 메서드 -> 양방향 연관관계에 추가해주면 좋음
-    public void addOrderItem(OrderItem orderItem) {
-        orderItems.add(orderItem);
-        orderItem.setOrder(this);
-    }
-
-    public List<OrderItem> getOrderItems() {
-        return orderItems;
-    }
-
-    public void setOrderItems(List<OrderItem> orderItems) {
-        this.orderItems = orderItems;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-
-    public LocalDateTime getOrderDate() {
-        return orderDate;
-    }
-
-    public void setOrderDate(LocalDateTime orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(OrderStatus status) {
-        this.status = status;
-    }
-}
