--- Eunji/week5/jpa-mapping/src/main/java/mapping/OrderStatus.java
@@ -1,5 +1,0 @@
-package mapping;
-
-public enum OrderStatus {
-    CANCEL, SUCCESS
-}
